% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{The log-sum-exp function}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{Is a vector of numeric values}
}
\value{
The log sum of exponentials
}
\description{
The log-sum-exp function
}
\examples{
##let x be a numeric vector
##x <- c(1:2000)
##log_summed_exps(x) = 2000.459

}
